:art: Refactor CSS-in-JS structure
