+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import SearchArea from './component/SearchArea/SearchArea';
-import { createGlobalStyle } from 'styled-components';
+import SearchArea from './component/SearchArea';
+import GlobalStyle from './App.styles';
 
 export default function App() {
   return (
@@ -10,15 +10,3 @@ export default function App() {
     </div>
   );
 }
-
-const GlobalStyle = createGlobalStyle`
-  @import url(//spoqa.github.io/spoqa-han-sans/css/SpoqaHanSansNeo.css);
-  * {
-    padding: 0;
-    margin: 0;
-    font-family: 'Spoqa Han Sans Neo', 'sans-serif';
-  }
-  body {
-    background-color: #cae9ff;
-  }
-`;
