+++ src/component/RecommendArea/Recommend.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
+import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import { RecommendStyled } from './RecommendArea.styles';
 import SearchIcon from '../../assets/icon_search.svg';
 
 const Recommend = ({ content }) => {
@@ -15,5 +15,15 @@ const Recommend = ({ content }) => {
 Recommend.propTypes = {
   content: PropTypes.string,
 };
+
+const RecommendStyled = styled.li`
+  svg {
+    margin-right: 12px;
+  }
+  display: flex;
+  align-items: center;
+  line-height: 2.6;
+  cursor: pointer;
+`;
 
 export default Recommend;
