+++ src/component/SearchArea.jsx
@@ -1,4 +1,31 @@
+import React from 'react';
 import styled from 'styled-components';
+import { useMediaQuery } from 'react-responsive';
+import RecommendArea from './RecommendArea/RecommendArea';
+import SearchIcon from '../assets/icon_search.svg';
+
+const SearchArea = () => {
+  const isPC = useMediaQuery({ query: '(min-width: 1040px)' });
+  return (
+    <SearchAreaStyled isPC={isPC}>
+      <Title className="title-responsive">
+        국내 모든 임상시험 검색하고
+        <br />
+        온라인으로 참여하기
+      </Title>
+      <SearchBar className="searchbar-responsive">
+        <SearchBar.Box>
+          {isPC && <SearchIcon />}
+          <input type="text" placeholder="질환명을 입력해 주세요. " />
+          {!isPC && <SearchIcon />}
+        </SearchBar.Box>
+        {isPC && <SearchBar.Button type="button">검색</SearchBar.Button>}
+      </SearchBar>
+      {/* show true/false에 따라 RecommendArea display 속성 변경 */}
+      <RecommendArea show />
+    </SearchAreaStyled>
+  );
+};
 
 const SearchAreaStyled = styled.div`
   padding-top: 80px;
@@ -21,7 +48,7 @@ const SearchAreaStyled = styled.div`
   }`}
 `;
 
-export const Title = styled.h1`
+const Title = styled.h1`
   text-align: center;
   font-size: 2.125rem;
   font-weight: 700;
@@ -30,7 +57,7 @@ export const Title = styled.h1`
   margin-bottom: 20px;
 `;
 
-export const SearchBar = styled.div`
+const SearchBar = styled.div`
   display: flex;
   flex-direction: row;
   background-color: #fff;
@@ -75,4 +102,4 @@ SearchBar.Button = styled.button`
   border-radius: 0 42px 42px 0;
 `;
 
-export default SearchAreaStyled;
+export default SearchArea;
